requirements.txt
python-dotenv==1.1.1
python-telegram-bot==13.15
redis==6.2.0
vk-api==11.9.9